requirements.txt
# requirements.txt
numpy
scipy
matplotlib
pandas
biopython
pydicom
Pillow
Flask